refactor (NormalRollingTest) : step2 - 투구 결과 객체 구조 변경에 대한 테스트 수정
