+++ src/main/java/bowling/domain/frame/State.java
@@ -1,6 +1,7 @@
 package bowling.domain.frame;
 
 public enum State {
+    READY,
     STRIKE,
     SPARE,
     MISS,
